wip: xdg: load icons, based on the "Icon"-entry in the .desktop files
First, search the main theme. This is done three times:
  - First, looking for an exact size match
  - Second, looking for closest size match
  - Third, picking any size

Number one may fail because there isn't *any* icons available in the
wanted size.

Number two may fail either for the same reason as number one, or
because the particular icon we're looking for isn't available in that
size.

Number three may fail if the icon we're looking for isn't available in
this theme.

If all three fails, we iterate the parent themes.