main: set optimal icon size from the font height
When searching for icons, we use the font height as-is. I.e. don't try
to find icons that are exactly the size of a whole line, as that looks
ugly.

However, when rendering, we *do* allow icons that are larger than the
font height, but still within the line height. This is to avoid
scaling.

When we do have to scale, we scale it to the font height, not the line
height.