main: report exec() failure back to parent process
Open a pipe before forking, with O_CLOEXEC set. If the exec()
succeeds, the pipe will be closed without anything having being
written to it.

If exec() fails, we write errno to the pipe, before exiting normally.

The parent process does a blocking read() on the pipe. If the exec()
succeeded, we'll get a 0-byte read. If it failed, we'll read the errno
value and log it.