main: don't commit a new surface update until previous commit has completed
When we want to draw a new frame, and the previous frame hasn't
completed yet, queue the new frame, and draw it at soon as the
previous frame has completed.

Note that we only keep the latest pending frame. I.e. if we manage to
trigger more than one update while still rendering the first frame,
only keep the last update.

This ensures a flicker-free experience :)