icon_load_theme: return a flat list of themes
Instead of returning a single "theme" instance, which contains a list
of themes it inherits, which in turn contain lists of themes they
inherit, return a flat list.

This way, we solve two problems; one, we ensure the themes are searched in
the correct order (if we use a tree structure, should it be searched
depth-first or breadth-first), and two, we can avoid search the same
theme multiple times (multiple themes may inherit the same theme(s),
at different levels).