main: hidpi: config values are absolute
That is, instead of scaling up everything, use the configuration
values as is.

This allows the user to, for example, configure a 1px border also on a
HiDPI screen.

Now, the surface width/height, must then be *divided* by the scaling
factor since normally, the surface geometry is absolute, and the rest
is scaled. We do the opposite.