main: read_cache(): optimize; application list is now sorted
Since both the application list *and* the cache file are now sorted,
we can optimize further, by progressing both lists in lock-step.

That is, for each cache line, we can start searching the application
list from where we stopped the last time. And we only need to search
until we find an application that either matches, or is "past" our
cache line.