render: render glyphs ourselves using freetype instead of cairo
Our font now handles font fallback when a glyph cannot be found in the
primary font (using fontconfig).

To be able to render this, we now render the glyphs ourselves using
pixman, instead of cairo.

Note: we could, probably, get rid of cairo completely, *if* we can
find a simple way to render the curved corners using pixman. The reset
is just boxes and rectangles, which can easily be done with pixman.