wayland: implement the surface listener protocol
Use this to detect actual scaling factor, and re-render.

If the user hasn't configured an explicit 'output', fuzzel will guess
the scaling factor to use. This could be wrong though (if the user has
more than one monitor, with different scaling factors).

Using the surface-enter event, we finally know, for sure, on which
output we were mapped. Thus we can use that output's scaling factor
and re-render.