wayland: wip: reload font at correct DPI when necessary
If we aren't able to correctly guess which output we actually end up
on, the DPI we used when rendering the first frame may be incorrect.

When this is the case, reload the font at the correct DPI,
re-calculate maximum number of matches and refresh.

This needs to be cleaned up.