wayland: guess initial subpixel mode
We didn’t initialize the subpixel mode at all, meaning it always
defaulted to FCFT_SUBPIXEL_DEFAULT, forcing a redraw if the output had
a different mode.

Now, we initialize it this way:

* If we have an explicitly configured monitor, use its subpixel mode
* If not, use the subpixel mode of the *first* monitor

The last point means we’ll always guess correctly if there’s only one
monitor. We may still guess wrong in a multi-monitor setup, in which
case the surface will be re-rendered immediately after being mapped.