main: flush text run caches on font reload
This fixes an issue where wrong font was being used for *some* entries
if we were guessing which monitor fuzzel would appear on, and guessed
wrong.

The reason being, of course, that we rendered one frame with the
“wrong” font, which set the text run cache on the items visible _on
that frame_. Then we reload the font, and bam, we’re using the old
text run cache for some entries.