svg: add support for using nanosvg as SVG backend
This patch removes the -Denable-svg=disabled|enabled|auto meson
command line option, and replaces it with a new one,
-Dsvg-backend=none|librsvg|nanosvg

Nanosvg is a tiny SVG parsing and rasterization library, bundled with
fuzzel. It doesn’t implement all available SVG features (for example,
it can’t do “text” elements). On the other hand, it doesn’t require
Cairo.

When the nanosvg backend is used, the rasterized version(s) of the SVG
image is generated on demand, and then cached. I.e. an SVG will be
rasterized the first time it is displayed.

Finally, -Denable-png=disabled|enabled|auto has been replaced with
-Dpng-backend=none|libpng, to match the new -Dsvg-backend option.