match: add support for paging
The matches object can now present multiple pages. This is completely
transparent to the caller.

Instead of limiting the internal counter ‘match_count’ to the maximum
number of simultaneously *visible* matches, make *all* items
accessible.

Since all users of the matches object uses accessor functions, we can
simply modify these to handle paging.

This patch also adds a couple of new functions, that the caller can
use to find out on which page we’re currently on, and the total number
of pages. This is currently not used, but can be used in the future to
e.g. display a "page n/m" hint.