match: add matches_selected_{prev,next}_page()
These functions move the current selection one page back, or one page
forward.

If we’re already on the first page, and we’re moving backward, the
selection is moved to the top.

If we’re already on the last page, and we’re moving forward, the
selection is moved to the last item.