wayland: frame perfection on multi-monitor setups
In a multi-monitor setup, fuzzel is forced to guess both which scaling
factor, and which DPI to use when rendering.

If we guess wrong, we’ll render one frame with “wrong” scaling and/or
DPI information, leading to a window of “wrong” size.

As soon as the surface has been mapped, we know the actual scaling
factor and DPI. If different from the guessed values, we correct
ourselves and re-render.

However, this is perceived as flicker. Most users *will* notice that
first, bad, frame.

https://codeberg.org/dnkl/fuzzel/issues/21 suggests adding a
--dpi-aware command line option to fuzzel. The idea being that if you
disable DPI awareness, we’re no longer depending on the DPI. I.e. the
font will have the same size regardless of DPI. Thus, the window size
will not change, even when we’ve guessed the DPI wrong.

That however is just a workaround. First, users may still want DPI
awareness. But worse, we’re still not handling bad scaling factor
values.

This patch takes a completely different approach. The idea is to get
mapped, but without actually displaying anything. Like before, we’ll
know the correct scaling factor and DPI values as soon as we’ve been
mapped, and can then correct ourselves.

How do we get mapped without displaying anything? We render a
completely transparent surface!

As a micro optimization, this -
render-one-initial-frame-completely-transparent - is only done when
actually necessary. I.e. when there are more than one monitors
available, *and* when the user hasn’t selected a specific
monitor (with the --output command line option).

Closes #21