icon: always search all XDG data directories when loading an icon
From
https://specifications.freedesktop.org/icon-theme-spec/icon-theme-spec-0.11.html

  In at least one of the theme directories there must be a file called
  index.theme that describes the theme. The first index.theme found
  while searching the base directories in order is used. This file
  describes the general attributes of the theme.

This is somewhat vague, but the way we interpret this is:

As long as we find an index.theme file in _some_
xdg-dir/icons/theme-name directory, _all_ xdg-dir/icons/theme-name
directories will be searched when loading an icon.

This means the per-user icons can be installed in
e.g. ~/.local/share/icons/hicolor, and still be found, even though
there’s no ~/.local/share/icons/hicolor/index.theme file, as long as
there’s an index.theme file somewhere
_else_ (e.g. /usr/share/icons/hicolor/index.theme).

This patch is actually quite simple; the icon_theme struct already
tracks the theme’s sub-directories as *relative*
paths (e.g. 128x128/apps/, or scalable/apps etc).

All we need to do is:

* Remove the ‘path’ member from the icon_theme struct
* Instead, loop through xdg_data_dirs()

The rest of the icon loading logic remains the same - as soon as we
find an exact match, use it - otherwise use the best match.

Note that the XDG data dirs are iterated in the outer loop, and the
theme dirs are iterated in the inner loop.

Closes #62