main: populate application list in a thread
Display the surface as soon as possible, to reduce the latency between
application startup and when we can start receive input.

The operation that takes the most time, by far, is when we populate
the application list, since this involves a lot of disk I/O.

Do this in a thread. When done, the thread signals this through an
event FD. The FDM handler for the event fd updates the matches object
and refreshes the GUI.

Closes #82