main: load icon themes, and icons, in a thread
First, this commit replaces all custom ‘context’ struct with a single,
super-context struct that contains everything needed by all callbacks.

Then, the icon loading logic is rewritten:

* Themes are loaded in the application populating thread, and only in
  non-dmenu mode

* Icons are also loaded in the application population thread, **if**
  we have an icon size (which we do after the wayland backend has called
  our font-reloaded callback at least once.)

Since we _also_ need to reload the icons in the font-reloaded
callback (when we guess wrong about which output we’ll get mapped on),
the icon reloading must be done in a critical section. If not, the
following could happen:

* The application population thread is started
* Wayland instantiates the font, and calls our font-reloaded cb
* The font-reloaded cb sets our context’s icon-size member
* The application population thread starts loading the icons
* Our window is finally mapped, but on an output whose DPI doesn’t
  matched the currently loaded font’s. The Wayland backed reloads the
  font and calls our cb
* Our cb reloads the icons **in parallel with the application
  population thread**

Chaos ensues.