main+render: protect icon loading+rendering with a lock
Now that icon loading is threaded, we may end up rendering at the same
time the icons are refreshing.

This is bad, since the renderer may access image data that has been
free:d.

To handle this, we introduce an “icon lock”. This icon lock is managed
by main(), and passed by pointer to the renderer.

This lock is taken before refreshing the icons (which happens either
when the font is being reloaded, or during startup, after we’ve
populated the application list).

When rendering, we *try* to take the lock. If we succeed, we render
all icons. If not, no icons are rendered.

The reason for this is: we don’t want to render icons of mixed
size. I.e. we don’t want a per-icon lock.

But we also don’t want to delay rendering (i.e. block on a mtx_lock()
call), since loading the icons can take some time, and during this
time, the GUI will appear to have frozen.

Note that we always refresh the GUI after having refreshed the
icons. Thus, if we do end up rendering a frame without icons, we’ll
re-render again, as soon as the icons have finished loading.