main: use a pipe instead of an event FD for thread signaling
The thread that populates the application list, and loads the initial
set of icons, previously used an event FD to signal when its done.

This was done by writing an event enum to the event FD; once when the
application list has been loaded, and once when the icons have been
loaded.

The problem is that an event FD will *add* the written value to its
current counter value.

In other words, if the thread finished very quickly, it’d end up
writing a ‘1’ and a ‘2’, *before* the main thread had had a chance to
read anything from the event FD.

Then, when the main thread *did* read the event FD, it’d get
‘3’ (1+2). This didn’t match any of the events, and nothing was
refreshed, resulting in an empty window.

This patch replaces the event FD with a regular pipe.