main: don’t unlink lock-file if we failed to acquire the lock
If we fail to acquire the lock (because another fuzzel instance is
running, and has the lock), don’t unlink the lock file when we exit.

This fixes an issue where running e.g.

  fuzzel&; fuzzel&; fuzzel

had only the second fuzzel instance refusing to run. Both the first,
and the third instance would run, since instance two removed the lock
file, making instance three succeed since it would create (and lock)
and file.