xdg: add applications with NoDisplay=true to our app list
Without doing this, a user .desktop file (with NoDisplay=true) will
not override the corresponding system entry.

It’s easy to see why: we never add the user .desktop file to our
application list (since NoDisplay=true). When we later find the system
.desktop file, we have no knowledge of the earlier user .desktop file,
and we end up (incorrectly) adding the system .desktop file to the
application list.

For this to work, we also need to add a new ‘visible’ member to the
application struct.

The dmenu backend has been updated to always set .visible=true, and
the match logic has been updated to filter out entries with
.visible=false.

Fixes #114