match: tokenize search text and match substrings individually
The search text is now split up on whitespaces, and each “token” is
search for individually.

An application is “matched”, if all search words are found. Search
words may be specified out-of-order.

Example: if the application title is “The Quick Brown Fox”, then
searching for “quick fox” will match, as will “fox quick”. But
searching for “quick red fox” will *not* match.

When classifying the match, the following rules apply:

- all tokens must match. That is, the application does *not* match, if
  at least one token fails to match.
- the application is an exact match *iff* there’s a single
  token, *and* that token is an exact match. Otherwise it’s a fuzzy
  match.