icon: regression: fix loading of icons from XDG_DATA_DIRS/pixmaps
In the early days, svg() and png() returned false if the specified
path either didn’t exist, or wasn’t a svg/png icon.

Since the icon loading refactoring, where icons are lazy-loaded, this
is no longer true: svg() and png() always returns true.

This broke the logic when trying icons in XDG_DATA_DIRS/pixmaps, which
relied on svg() and png() failing if the file does not exist.

This patch fixes this, by explicitly checking for the file’s
existence, using faccessat(), before “loading” the icon.