main: ignore -i
The -i flag is used in other, similar utilities, to mean “case
insensitive search”. Let’s parse it, but ignore it, for compatibility
with said other utilities.

Note that this means you can no longer set the icon theme with “-i
THEME”. Use the long option (“--icon-theme=THEME”) instead, or set it
in the configuration file.

Closes #149