config: add exit-on-keyboard-focus-loss option, defaults to true
When set to false, or the corresponding
--no-exit-on-keyboard-focus-loss command line option is used, fuzzel
will not exit when it receives a keyboard leave event.

This is useful on compositors where enabling their focus-follows-mouse
option causes fuzzel to exit as soon as the mouse is moved (assuming
mouse is not initially over fuzzel), or as soon as the mouse leaves
the fuzzel window.

Closes #128