xdg: use localized strings, when possible
Use the locale set in LC_MESSAGES, and map against the .desktop files
localized keys.

According to the XDG spec, locales should be matched in this order:

* lang_country@MODIFIER
* lang_country
* lang@MODIFIER
* lang
* (default - non-localized)

To avoid having to scan all keys multiple times, we keep track of each
key’s “locale score”. The default score is 1 (the non-localized key),
“lang” has score 2, and lang_country@MODIFIER has score 5.

When parsing the key, extract the locale and match against
LC_MESSAGES, in the order mentioned above. Also remember the score.

Then, when matching the key, also check that key’s locale score, and
only update it if the current score is higher than the last one.

Closes #199