icon: fix loading of absolute-path PNG icons
Before we implemented lazy-loading of icons, the svg() function
returned false if the icon failed to load (for example, if it’s not an
SVG, but e.g. a PNG).

Now, both svg() and png() always returns true. This broke loading of
PNG icons with an absolute path, since we always tried svg() first.

Fix by checking the file extension, and choosing svg() or png()
accordingly.

Closes #214