wayland (regression): don’t dereference ‘app’ when it’s NULL
When requesting an XDG activation token, we need the app ID. This is
stored in the ‘application’ struct, and thus we were dereferencing our
‘app’ pointer to get it.

The problem is ‘app’ may be NULL in this context. For example, when
executing a command line (instead of a match entry).

Fix by simply skipping the activation token.

Closes #222