render: remove check that skips rendering lines if we’re outside the buffer
First, the logic doesn’t always work; the user may have configured a
smaller line-height, than that of the font. The check would always
trigger, and the last element always skipped.

Second, we’re already setting a clip region on the buffer, so we’re
already protected against writing outside the pixman buffer.

Thus, it makes sense to render the (typically last) line, even if it
may get truncated.

Closes #234