wayland: store output size in inches as floats
The struct monitor’s ppi field values are calculated from the monitor
dimensions in inches, which are in turn calculated from the ones in
millimetres. Before this patch all those values were integers, so there
was a noticeable error in these calculations; for example, in case of a
1920×1080 Dell E2418HN panel:
 - the original dimensions in millimetres are 527.04 × 296.46,
 - let’s assume Wayland rounds them to 527 × 297,
 - the dimensions in inches are ca. 20.79 × 11.69 (20.75 × 11.67 in real
   world), but they stored as integers, so actually they are 20 × 11,
 - the horizontal and vertical PPIs are 96 × 98.18, so 96 × 98 as
   integers, while the real world panel has 92.53 in both directions.

If the dimensions in inches were stored as floats, which is the aim of
this patch, the horizontal and vertical PPIs would be 92.54 × 92.36,
which, even as integers, are much closer to the real values.

The differences get worse if Wayland reports wrong size of the panel:
for example, I’ve got a monitor which PPI is calculated, without and
with this patch, respectively, as 96 × 100 and 93 × 92, while it
actually is 94.

The calculated PPI values don’t seem used anywhere except in a LOG_INFO
call. A value that differs too much from the expected one may cause some
users to write a patch like this one.