application: don't ignore empty arguments
Before this patch, a desktop entry with

    Exec=foo "" bar

would result in the following argv being executed: ["foo", "bar"].

This is wrong, and the reason is tokenize_cmdline(), before this
patch, required an argument to be at least one character long.

This patch fixes this by special casing quotes (more than before); we
rely on the part of the specification that says that arguments have to
be quoted in whole. This means quoted arguments that aren't preceded
and followed by space characters are invalid.

Instead of memmoving the string to "erase" quotes, we simply reset
search_start to point to the character *after* the opening quote, and
then unconditionally push an argument when the closing quote is
detected. After this, search_start is set to the character after the
closing quote.

Result:

* empty (quoted) arguments are now recognized, and kept.
* foo bar"zoo" will result in 'foo zoo' being executed (bar"zoo" is
  invalid, and fuzzel ignores the characters preceding the opening
  quote).
* foo "bar"zoo will result in 'foo bar zoo' being executed ("bar"zoo
  is invalid, and fuzzel treats the characters following the closing
  quote as a separate argument)

Closes #285