xdg: fix crash when parsing line that begins with whitespace
We strip trailing whitespaces by moving the 'line' pointer forward. In
case the whitespace trimming results in an empty string, we free the
line data and then continue with the next line in the .desktop file.

This results in a crash in free(), since the line pointer is no longer
pointing to the beginning of the allocated line data.

Fix by maintaining two line pointers; one that always points to the
beginning of the allocated data (the 'raw' data), and one that points
to the line data that we'll later parse.

Closes #328
Closes #329