match: improve sorting logic with --no-fuzzy
Score matches, and sort matches with a higher score before matches
with a lower score.

The score is equal to the longest consecutive match string. So, for
example, when searching for 'oo', 'foot' will sort before 'firefox
browser'; the longest match string in 'foot' is 2, while in 'firefox
browser' it's 1.