wayland: do the first-frame-is-transparent trick when fractional scaling is available
When fractional scaling is enabled, we have no way to *guess* the
correct scaling factor. We *must* wait for the compositor to report
it.

Unfortunately, all(?), or at least most, compositors currently report
the scaling factor *after* mapping the window.

This leads to flicker if fractional scaling is used.

Fix by rendering the first frame fully transparent when fractional
scaling is available.