render: make sure scale is updated on scale changes
Before this change, render_set_font() was only called when the *font*
had been reloaded. With dpi-aware=yes, that doesn't happen on scale
changes.

This caused the cached scale value (in the renderer) to be
wrong. This meant mouse selection didn't work correctly, since the
logic that calculates which row the cursor is on depends on
the *render's* scale value be correct.

Another issue was that UI elements that _do_ depend on the scale
factor also wasn't updated.