wayland: don't set hover selection on entry, or first motion
When starting fuzzel, the compositor will send a pointer enter event
if the mouse is over the newly appeared fuzzel window. This caused us
to select an entry in the middle of the list, instead of the first
entry.

But, disabling hover selection on pointer entry isn't enough. We may
also get a motion event, if we change the scaling factor of the
window (the compositor does this to ensure we have the correct cursor
position, for the new logical coordinate space).

So,

1. don't do a hover selection on pointer entry
2. don't set pointer.{x,y} on entry, since that allows us to...
3. don't do a hover selection on pointer motion if pointer.{x,y} are 0

In short, we only do a hover selection after receiving *two* motion
events.

This appears to be good enough to prevent the synthetic events from
triggering a hover selection.