match: performance: cache lower-case strings + incremental matching
This patch greatly improves the performance of the matching logic, by
doing essentially two things:

1. avoid having to call c32len() and toc32lower() over and over and
   over again.
2. When *adding* characters to the search string, do an incremental
   match where we only match against already matched entries, instead of
   matching against all available entries.

A perf recorded of fuzzel --dmenu with 80K entries showed toc32lower()
an c32len() used *a lot* of CPU. This shouldn't come as a surprise; we
do case insensitive matching by calling toc32lower() and both the
search string, and the searched strings.

However, we were doing this while matching. That is, over and over and
over again...

Similarly, c32len() was called a lot on the searched strings.

This patch caches the lower cased strings from the application
metadata so that we only need to call toc32lower() when loading the
.desktop files (or dmenu items, for that matter). It also converts the
search string to lower case *once* each time we perform a match.

Next, when *adding* characters to the search string, there's no need
to search *all* applications again. We can limit the search to already
matched items. The exception is when we're doing fuzzy matching, where
we also need to take the minimum fuzzy match length into account.