wayland: performance: render frames on-demand, rather than eagrly
Before this patch, frames were rendered eagerly. If we were still
waiting for a frame callback, we stashed the rendered frame as
"pending". If we rendered even more frames before the frame callback,
those frames replaced the "pending" frame.

Once we got the frame callback, we simply committed the pending frame.

This patch changes the logic to rendering frames on-demand. That is,
if we're waiting for a frame callback, do *not* render a frame at
all. Instead, just remember that we need to render a new frame, and
then once we get the frame callback, render and commit it.

This ensures we don't spend CPU cycles on rendering frames that we
just end up throwing away.

Now, this wasn't _that_ much of a problem, but was observable with
e.g. high keyboard repeat rates.

The downside of this patch is that it *potentially* increases
latency; *if* we're slow at rendering a frame, waiting for the frame
callback before even starting to render the frame _could_ cause us to
miss a display cycle, increasing lag with ~16ms (for 60HZ).