render: render match list using multiple rendering threads
The match list is what takes the longest time to render. Both due to
the fact that it's a lot of text (comparatively), and all the icons.

This patch adds rendering threads.

1. Render background on main thread
2. Render prompt on main thread
3. Render match list on the rendering threads

This patch also "bumps" the SHM boilerplate to the latest one from
foot. This brings better buffer caching. This will allow us to improve
rendering performance even more in the future. For example, by
skipping the background if we're re-using an old buffer.

The user can control how many threads fuzzel uses with:

* 'workers' option in 'fuzzel.ini'
* The '--workers' command line option