main: --no-run-if-empty: display window as soon as we see _some_ input
This allows us to have 'exit-immediately-if-empty' enabled at all
times, while still displaying a window immediately once there's
anything in STDIN. That is, we no longer hold off displaying the
window until STDIN has been closed, even with --no-run-if-empty.

This is implemented by always loading dmenu entries in a thread. This
is better, since we now use a single code path for dmenu loading.