match: sorting: always prefer 'exact' matches over 'fuzzy' matches
The only time a match is 'fuzzy' is when it's a levenshtein
match (i.e. we only ever see fuzzy matches with --match-mode=fuzzy).

'Exact' on the other hand means we matched a *substring* exactly (or
using fzf).

So, to summarize, depending on --match-mode, we see the following
types of matches:

* --match-mode=exact: exact
* --match-mode=fzf: exact
* --match-mode=fuzzy: exact, fuzzy

Before this patch, the match type was ignored when sorting the
entries. This meant a fuzzy match, like 'bulg', would sort higher than
an exact match like 'bug' (assuming what you typed was 'bug). The
reason is simple; the fuzzy match is longer, thus giving it a higher
match score than the exact match.

This patch changes the sorting logic to always prefer exact matches
over fuzzy matches. Since we never get any 'fuzzy' matches with
--match-mode=exact|fzf, this patch only affects --match-mode=fuzzy.