wayland: ignore size hints from compositor
We set the size ourselves. Fixes a crash when changing the scaling
factor at run-time; the compositor would send a configure event with a
logical pixel size that matched the old size, causing us to re-set our
width/height to something that doesn't match the new font size we're
using after the scale change. That in turn lead to pixman operations
going outside the pixman buffer, causing fuzzel to crash.