config: don't ignore border.radius
It used to be that we required cairo for rounded corners. So, we
detected when the user tried to use rounded corners in non-cairo
builds, and warned them it wasn't supported. We also ignored the
value.

None of this is true anymore - we can do rounded corners just fine in
non-cairo builds. So, remove the warning, and don't ignore the value
set by the user.