render: don't pass a negative radius to fill_rounded_rectangle()
If the border *size* is bigger than the border *radius*, we ended up
passing a negative radius to fill_rounded_rectangle().

Since the radius parameter in fill_rounded_rectangle() is an uint16_t,
we ended up passing a *very large* value, e.g. 65535. This resulted in
an integer overflow when calculating the number of pixman rectangles
to use, leading to a crash when calling
pixman_image_fill_rectangles().

Closes #401