icon: skip 'scaled' icon theme directories
These directories contain icons that are physically larger (more
pixels), but with the same level of detail as unscaled directories.

For example, a 16x16@2 directory may contain 32x32 icons, but with the
same level of detail as a 16x16 directory.

Up until now we've been matching these using the scaled pixel
size (i.e. 32x32, in the example above).

This means we've been, for some font/icon sizes, been picking icons
with less details than necessary, because we've been ignoring the
scaling factor of the target output.

Unfortunately, the scaling factors in icon themes are integers, with
the smallest possible value (besides 1) is 2. Nearly no-one uses that
much scaling.

So, instead of trying to bring the output scaling factor into the
game (which, when rounded to an integer, will almost always be 1
anyway), simply ignored all 'scaled' icon directories.

Closes #459