icon: search $HOME/.icons, and fix how we search for standalone icons
$HOME/.icons should be searched before any of the other XDG data dirs.

Also update how we look for standalone (non-theme) icons. Before this
patch, we looked for icons in the pixmaps sub-directory, of each XDG
data dir.

According to
https://specifications.freedesktop.org/icon-theme-spec/latest/#directory_layout,
this is incorrect; we should only search /usr/share/pixmaps.

Furthermore, we should _also_ look for standalone in *all* directories
we look for icon themes in.

In other words, the directories we look for icons and icon themes
are (in this order):

* $HOME/.icons
* XDG_DATA_DIRS/icons
* /usr/share/pixmaps

All directories are first searched for an icon in the selected icon
theme. If there is no themed icon in any of the above directories, we
once again loop the directories, this time looking for a standalone
icon.