dmenu: --with-nth: don't match "other" columns
When using --with-nth, fuzzel still matched the entire input
string. This is not what e.g. fzf does. It also caused the match
highlighting to be offset.

Fix by parsing out the column to be displayed when loading the dmenu
entries, rather than when rendering them. Store the full dmenu input
in a separate member (not used by the match logic) in the application
struct.

Closes #551