config: add scaling-filter option
Before: this patch, lanczos3 was always used (unless the PNG was >=
1024px, in which case we used pixman's "fast" filter). Now, the
default is 'box'. This filter is much faster than lanczos3 (and
lanczos2 and cubic, though not by as much), while still looking just
as good (in most cases at least).

The check for too large icons remains, but has been slightly modified:

If the icon is >=1024px, we switch to the fast filter, regardless of
which filter the user has selected.

If the icon is >=256px (but less than 1024), *and* if the user
selected a "slow" filter (we consider cubic, linear and lanczos* to be
slow), then warn and use the box filter instead.

Closes #553