render: use the width of 'o' instead of 'W' when calculating window width
This gives a better estimate when a variable width font is used.

'W' is one of the widest characters (in most fonts at least...)
meaning we almost always were able to fit close to twice as many
characters, than specified in --width.

Also update the documentation of '--width', to mention that it is an
estimate when a variable width font is used.

Closes #560