Add new --minimal-flag option
Useful for building small dmenu menus with a dynamic number of lines
that's less than --lines.

Fuzzel prevents displaying anything until all dmenu entries have been
loaded, at which point the number of display lines to use are
calculated (the minimum of --lines, and the total number of dmenu
entries).

Co-authored-by: Daniel Eklöf <daniel@ekloef.se>