✨ Add {cmd} in --terminal command, like Rofi. (#600)
{cmd} may appear more than once. If it appears at all,
it it not also appended.

Fixes #564

## QA Log

Tested with:

```
/pkg/fuzzel/usr/bin/fuzzel --config=test_terminal.ini
```

Where test_terminal.ini contains:

```
[main]
terminal=foot -a '{cmd}' -T '{cmd}' {cmd}
```

Then I checked the title and app_id to confirm that they were both set to the command.

For terminals like foot that allow setting the app_id and title, this is a nicer default.

Reviewed-on: https://codeberg.org/dnkl/fuzzel/pulls/600
Co-authored-by: Mark Stosberg <mark@stosberg.com>
Co-committed-by: Mark Stosberg <mark@stosberg.com>