✨ New Feature: Hide Input Line (#597)
Hiding the prompt/input line may be of interest for a more minimal design, and also makes Fuzzel for building small menus or touch-based menus where a filtering feature is not necessary.

To select items with the prompt hidden, arrow keys can be used or... as a hidden feature you can still type input as before, but it won't be visible.

Fixes #430
Replaces #596

## Questions

 * Is --hide-prompt clear enough? Perhaps that sounds like it refers to just the prompt and not the whole line. Alternate suggestion: `--hide-input-line`. Including "line" in the name make sit clearer that the whole line is being hidden, not just the  prompt part.

If this feature is approved, there's still a bit more clean-up work to do with docs/completions/changelog.

## Screenshot

![image](/attachments/d09350ee-49d5-4346-8861-35174a9228f9)

That was made with:

```
echo -e '⏻Power Off\n󰜉 Restart' | ./pkg/fuzzel/usr/bin/fuzzel --hide-prompt --minimal-lines --dmenu --width 10
```

The version includes the UX idea of still allowing input, although it's now hidden. That means in addition to selecting with the arrow keys or via touch, you can still filter as before!

If you normally use Fuzzel by typing a few characters and hitting enter, this offers an option for an even more minimal aesthetic.

Reviewed-on: https://codeberg.org/dnkl/fuzzel/pulls/597
Reviewed-by: dnkl <dnkl@noreply.codeberg.org>
Co-authored-by: Mark Stosberg <mark@stosberg.com>
Co-committed-by: Mark Stosberg <mark@stosberg.com>